requirements.txt
django>=1.9.1
django-model-utils>=2.4
xmltodict>=0.9.2
M2Crypto>=0.23.0
